refactor(modules): initial split of rpm.rs
